# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Stalk market prediction made programmatic"
  version: "1.0.0"
  title: "Stalks as a Service"
  contact:
    email: "you@your-company.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/highb/animal-crossing-as-a-service/1.0.0"
schemes:
- "https"
- "http"
paths:
  /stalkPrediction:
    get:
      summary: "Stalk Market Prediction"
      description: "By passing in the appropriate options, you can get all predictions\
        \ for your islands Stalk Market\n"
      operationId: "stalkPrediction"
      produces:
      - "application/json"
      parameters:
      - name: "firstTimeBuyer"
        in: "query"
        description: "Is this your first time buying turnips from Daisy Mae on your\
          \ island?"
        required: false
        type: "boolean"
        default: false
      - name: "previousPattern"
        in: "query"
        description: "What was last week's turnip price pattern"
        required: false
        type: "string"
        enum:
        - "LargeSpike"
        - "SmallSpike"
        - "Fluctuating"
        - "Decreasing"
      - name: "sellPrice"
        in: "query"
        description: "How much Turnips are selling for this week"
        required: false
        type: "integer"
        maximum: 999
        minimum: 0
        format: "integer"
      - name: "buyPrice"
        in: "query"
        description: "The daily prices that the Nooks are willing to pay in order\
          \ from Monday AM/PM to Saturday AM/PM prices."
        required: false
        type: "array"
        items:
          type: "integer"
        collectionFormat: "multi"
        maxItems: 12
        minItems: 0
      responses:
        "200":
          description: "Daily predictions for Stalk Market prices"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Prediction"
        "400":
          description: "bad input parameter"
      x-swagger-router-controller: "Default"
definitions:
  Prediction:
    type: "object"
    required:
    - "maximumOverall"
    - "maximumPrices"
    - "minimumOverall"
    - "minimumPrices"
    - "pattern"
    - "probability"
    - "sellPrice"
    properties:
      pattern:
        type: "string"
        enum:
        - "LargeSpike"
        - "SmallSpike"
        - "Fluctuating"
        - "Decreasing"
      probability:
        type: "number"
        example: 20
      sellPrice:
        type: "integer"
        example: 151
      minimumPrices:
        type: "array"
        example:
        - 128
        - 120
        - 113
        - 105
        - 98
        - 90
        - 83
        - 135
        - 211
        - 302
        - 211
        - 135
        items:
          type: "integer"
        maxItems: 12
        minItems: 12
      maximumPrices:
        type: "array"
        example:
        - 136
        - 132
        - 127
        - 123
        - 118
        - 114
        - 109
        - 212
        - 302
        - 906
        - 302
        - 212
        items:
          type: "integer"
        maxItems: 12
        minItems: 12
      minimumOverall:
        type: "integer"
        example: 302
      maximumOverall:
        type: "integer"
        example: 906
    example:
      minimumPrices:
      - 128
      - 120
      - 113
      - 105
      - 98
      - 90
      - 83
      - 135
      - 211
      - 302
      - 211
      - 135
      maximumOverall: 906
      probability: 20
      maximumPrices:
      - 136
      - 132
      - 127
      - 123
      - 118
      - 114
      - 109
      - 212
      - 302
      - 906
      - 302
      - 212
      pattern: "LargeSpike"
      minimumOverall: 302
      sellPrice: 151